[RacingCar] refactor : GameManager
 접근 제한자 변경
