+++ src/test/java/view/ValidateInputTest.java
@@ -1,4 +1,4 @@
-package controller;
+package view;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
